Abstract common logic of aggregate operators
