+++ sql/core/src/main/scala/org/apache/spark/sql/execution/aggregate/SortAggregateExec.scala
@@ -24,7 +24,7 @@ import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.expressions.aggregate._
 import org.apache.spark.sql.catalyst.plans.physical._
 import org.apache.spark.sql.catalyst.util.truncatedString
-import org.apache.spark.sql.execution.{AliasAwareOutputPartitioning, ExplainUtils, SparkPlan, UnaryExecNode}
+import org.apache.spark.sql.execution.{AliasAwareOutputPartitioning, SparkPlan}
 import org.apache.spark.sql.execution.metric.SQLMetrics
 
 /**
@@ -38,7 +38,8 @@ case class SortAggregateExec(
     initialInputBufferOffset: Int,
     resultExpressions: Seq[NamedExpression],
     child: SparkPlan)
-  extends UnaryExecNode with AliasAwareOutputPartitioning {
+  extends AggregateExec(groupingExpressions, aggregateExpressions)
+    with AliasAwareOutputPartitioning {
 
   private[this] val aggregateBufferAttributes = {
     aggregateExpressions.flatMap(_.aggregateFunction.aggBufferAttributes)
@@ -110,22 +111,6 @@ case class SortAggregateExec(
   override def simpleString(maxFields: Int): String = toString(verbose = false, maxFields)
 
   override def verboseString(maxFields: Int): String = toString(verbose = true, maxFields)
-
-  override def verboseStringWithOperatorId(): String = {
-    val allAggregateExpressions = aggregateExpressions
-
-    val keyString = groupingExpressions.mkString("[", ", ", "]")
-    val functionString = allAggregateExpressions.mkString("[", ", ", "]")
-    val inputString = child.output.mkString("[", ", ", "]")
-    val outputString = output.mkString("[", ", ", "]")
-    s"""
-       |(${ExplainUtils.getOpId(this)}) $nodeName ${ExplainUtils.getCodegenId(this)}
-       |Input: $inputString
-       |Output: $outputString
-       |Keys: $keyString
-       |Functions: $functionString
-     """.stripMargin
-  }
 
   private def toString(verbose: Boolean, maxFields: Int): String = {
     val allAggregateExpressions = aggregateExpressions
